Seven Sages: clear every narrowing conversion on the branch's changed lines
MSVC builds soh/ at /W3 /WX, so C4244 is fatal. The previous pass only looked at
SevenSages/ and EnrichWorld/ and triaged what it found; this one scans every file
the branch changes against develop and fixes all of it.

SevenSagesSongMagic: the release-watch target is now s8, matching gSaveContext.magic
where it gets written back, instead of narrowing at both write sites. magicTarget is
declared s16 but Magic_RequestChange range-checks it against 0 and magicCapacity
before returning true, so it always fits.

The two soh/src/ files are ours as well - src/*.c is compiled /w on MSVC so they
could not have broken the build, but the added lines were narrowing all the same.

+++ soh/soh/Enhancements/SevenSages/SevenSagesSageCosmetics.cpp
@@ -137,7 +137,7 @@ uint8_t MinChannel(Rgb c) {
  */
 Rgb TowardBlack(Rgb c, float t, int32_t minDelta = 0) {
     uint8_t room = MaxChannel(c);
-    if (minDelta > 0 && room > 0 && room * t < minDelta) {
+    if (minDelta > 0 && room > 0 && room * t < static_cast<float>(minDelta)) {
         t = std::min(1.0f, static_cast<float>(minDelta) / room);
     }
     return { ClampChannel(std::lround(c.r - c.r * t)), ClampChannel(std::lround(c.g - c.g * t)),
@@ -146,8 +146,8 @@ Rgb TowardBlack(Rgb c, float t, int32_t minDelta = 0) {
 
 Rgb TowardWhite(Rgb c, float t, int32_t minDelta = 0) {
     int32_t room = 255 - MinChannel(c);
-    if (minDelta > 0 && room > 0 && room * t < minDelta) {
-        t = std::min(1.0f, static_cast<float>(minDelta) / room);
+    if (minDelta > 0 && room > 0 && static_cast<float>(room) * t < static_cast<float>(minDelta)) {
+        t = std::min(1.0f, static_cast<float>(minDelta) / static_cast<float>(room));
     }
     return { ClampChannel(std::lround(c.r + (255 - c.r) * t)), ClampChannel(std::lround(c.g + (255 - c.g) * t)),
              ClampChannel(std::lround(c.b + (255 - c.b) * t)) };
